requirements.txt
fastapi
uvicorn
streamlit
pydantic
faiss-cpu
sentence-transformers
pypdf
python-dotenv
google-generativeai
